refactor: move submit to inside component to avoid @ts-expect-error
